-bitroad.net: RIP fixes #7294 -improved soundcloud errorhandling fixes #7295 -improved peeje errorhandling fixes #7297 -improved gigasize errorhandling fixes #7298 -fixed xxxaporn fixes #7301 -fixed debridlink expire time -fixed sockshare premium -fixed depositfiles...wtf -added support for dollyshare -improved qq decrypter refs #6193 -improved mediafire errorhandling
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@19460 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/plugins/hoster/DebridLinkFr.java
@@ -103,7 +103,7 @@ public class DebridLinkFr extends PluginForHost {
         case 1:
             ac.setStatus("Premium Account");
             account.setValid(true);
-            ac.setValidUntil(System.currentTimeMillis() + Integer.parseInt(getJson("PREMIUMLEFT")) * 1000);
+            ac.setValidUntil(System.currentTimeMillis() + Long.parseLong(getJson("PREMIUMLEFT")) * 1000);
         case 2:
             ac.setStatus("Lifetime Premium Account");
             account.setValid(true);
